qwen35, qwen35moe: one tokens_per_second, not two - plus the 21 warnings that compile error was hiding
`crates/qwen35/tests/vl.rs` and `crates/qwen35moe/tests/vl.rs` each set
`tokens_per_second: 2` TWICE in one `VisionConfig` literal, so both test
binaries failed to compile (E0062) on HEAD. 61ed68b16 and fa2425f94 each added
that missing field independently, at different lines of the same literal, and
two insertions at different offsets in one file are a textually clean merge -
nothing had anything to report until rustc did.

That error aborted the clippy gate before it reached six other crates. The gate
said so rather than reporting a number it could not stand behind, and behind the
abort were 21 pre-existing warnings against a baseline of 0. Fixed here rather
than absorbed into the baseline, since that baseline is a floor and not a
ratchet with slack in it:

- `gguf::kquant`: `try_kq_rect`'s success value becomes a named `KqRect`. Three
  of its four members are `Vec<u32>`, so the bare tuple said nothing at a call
  site about which plane is which; the alias carries that. One `g % 2 == 0`
  becomes `g.is_multiple_of(2)`.
- `qwen3tts::caps`, `qwen3tts::pipeline`, `qwen3tts`'s `mtp` test and
  `qwen3::serve`: `doc_lazy_continuation`, and the RENDERED doc was wrong, not
  just the lint. A wrapped prose line that happens to begin with `- ` is a
  CommonMark list marker, so every line after it was parsed as a continuation
  of a list item nobody wrote. Reflowed so no line opens with one; the prose is
  unchanged.
- `qwen3tts`'s `asr_roundtrip` WER helper walks its own row instead of a range;
  `qwen35::int8_gguf_resident` had a `vec!` that never grows; `glmdsa::sample`
  a `u32 as u32`.

`make clippy` is now exit 0, 0 warnings.

Recorded as lessons.md #88, because neither half is specific to these two files:
"add the one missing field" is exactly the fix two parallel sessions produce
independently and merge cleanly into a duplicate, and a gate that aborts leaves
an UNKNOWN warning count behind it, not a zero one.

+++ crates/glmdsa/src/sample.rs
@@ -175,7 +175,7 @@ mod kv_gen_tests {
         let cfg = GlmConfig::tiny();
         let init = crate::init::init_weights(&cfg, 11);
         let model = Glm::new_on(gpu_core::testgpu::dev(crate::model::PIPELINES), cfg.clone(), 1, 64, &init);
-        let vocab = cfg.vocab as u32;
+        let vocab = cfg.vocab;
         let prompt = vec![2u32, 6, 4, 1, 7, 3, 5];
         let max_new = 32;
